BUG-2350: do encapsulte null snapshot
getSnapshot() has only one caller in the same package, so performing
Optional.formNullable() just so we can then check for presence means
unnecessary object creation. Remove it and just deal with the potential
nulls.

Also rename methods for clarity and do not allow direct stores of nulls
into the cache.

Change-Id: I671a075ad6200fc37e586e413256bdd57e9ae4a0
Signed-off-by: Robert Varga <rovarga@cisco.com>