Set root pom.xml <name> for Sonar
As mentioned on the mailing list Sonar uses the <name> field of the
pom.xml that is passed to the mvn command as the name of the project in
Sonar. In most cases this is the root pom.xml file in a project. This
patch sets the name to the project shortname. In this case "controller".

https://lists.opendaylight.org/pipermail/discuss/2014-November/004024.html

Change-Id: I52a1cc89da77913da6e530b2d2776538887c9b63
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>