Update patch set 1
Patch Set 1:

We'll be removing the extensions from yang files as suggested. We'll add a yang file in restconf doc explorer project that will contain 2 properties

1. List of modules to hide
2. List of augmentations to hide data nodes

With this the existing yang files will not be edited. Also modules to show/hide will be specific to restconf api explorer.

What do you think?

Patch-set: 1