BUG-2383: deprecate ModificationType.MERGE
The entire ModificationType enum is an accidental leak from the
implementation. The most pressing deficiency is that we do not want to
leak the merge operation, but report it either as a subtree modification
or a write.

Introduce a new type, LogicalOperation, which contains all the current
and use it internally. The new name also implies slight changes in how
the states are called:

ModificationType.UNMODIFIED maps to LogicalOperation.NONE
ModificationType.SUBTREE_MODIFIED maps to LogicalOperation.TOUCH

Change-Id: I446e4391583de060b2ab722deb5c004f60c8e6e7
Signed-off-by: Robert Varga <rovarga@cisco.com>