BUG-2876: improve ModificationType accuracy
Application of an empty merge into a container results in no
modification of that container. From physical (ModificationType)
perspective, this means we can report UNMODIFIED allowing users
to skip one level of processing.

Also mark follow-up work for improving modification pruning at the cost
of performing a search in the children map.

Change-Id: I9c1c9509d3f6234570892bd90232fca3d76bd2ba
Signed-off-by: Robert Varga <rovarga@cisco.com>