Update patch set 4
Patch Set 4:

I am going to put the race/safety issue to a temporary rest (I would still like to address it as a part of looking at better alternatives to byte []) but not now.

I will do this by issuing a Rel 5 which has an extra constructor which will clone the byte[] and an extra getter which will allow you to get a cloned byte [] out.

Patch-set: 4