BUG-2939: fixup dynamic nodes to cover overwrites
Promote inner classes to their proper equivalents. Perform proper lazy
comparison to cover all possible scenarios where the user might
want to examine terminal nodes in greater depth. This is done by
tracking presence of before/after data and fudging appropriate
modification types.

We make sure to not run equals() on the data, as that may introduce very
steep costs -- an overwrite with the same value is still reported as a
full write, not as unmodified.

Change-Id: I0df100a74cfa82ceddeefcfae0ca2a011c608e08
Signed-off-by: Robert Varga <rovarga@cisco.com>