Update patch set 4
Patch Set 4:

I have an updated version which does not use shell to do overrides and has v4 address. However, git (quite stubbornly) refuses to update once more an existing commit in my tree to resubmit this as a new version the same gerrit. This is probably related to something to do with me having added and deleted files one time to many. 

I am going to submit that as a new gerrit shortly (and update that with v6 once that is done - probably tomorrow). I will add the same list reviewers for it too. Once that is submitted, I will mark this one as abandoned.

Patch-set: 4