Bug 2825 - Proposal for human override of generated code
This implements:

    1. Override for MacAddress
    2. Override for Ipv4Address
    3. Override for Ipv6Address

Version 7 - fix for javadocs, v6.

Taking into account all suggestions so far - encoding, guava, etc

Note - with the current state of v6 in the controller once the v6
change is in, all hell will break loose. However, it is for the
better as it will break only where it is non-compliant or broken
at present.

Change-Id: I5a0adb4fc973f25770bf11e081985c0d6ae181e6
Signed-off-by: Anton Ivanov <aivanov@brocade.com>