WIP: Proposal for human override of generated code
This is a plain replacement of generated code, which does not really
work.

What it attempts to do is to codify an in-memory representation which
is normalized to canonical format based on description.

The proper approach is to allow NormalizedNode deserialization logic
(i.e. yang-data-codec-*) to understand that:
  1) there is a canonical format (human-readable in description)
  2) that canonical format is *not* the same as what is on-wire
  3) there is a more efficient storage than java.lang.String

All these points are addresed via yang.common.CanonicalValue and
related interfaces. What we are lacking is the actual plugin architecture
to add user-defined CanonicalValues and have codecs integrated with it.

We are keeping this patch for reference of how things should look like
for ietf-{inet,yang}-types in the end.

JIRA: YANGTOOLS-418
Change-Id: I5a0adb4fc973f25770bf11e081985c0d6ae181e6
Signed-off-by: Anton Ivanov <aivanov@brocade.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>