Update patch set 2
Patch Set 2:

I noted that the osgi bundle fails on sun.net.util. I am open to suggestions on what to replace the parser from there with. 

Just to be clear - guava does not work. Passing non-canonical v6 addresses into Guava results in an exception (this also means that quite a bit of the v6 functionality in ODL at present is broken).

I am going to do some digging, but so far I have been able to find only this class to do what we need.

Patch-set: 2
Label: Verified=0