Bug 3344: StackedPathArguments did not expected empty stack.
StackedPathArguments.Iterator used only stack iterator
to check for next value, but did not account for possibility
that stack is empty.

This was possible if StackedYangInstanceIdentifier was accessed
from two threads, both entered null checked section and both
tried to create stack. The slower one created StackedPathArguments
from original StackedPathArguments as it is shown in Unit test.

Added unit test which replicates situation from live-run
and tests if it works.

Change-Id: I97dc4d1a6e7b38206b475a7d1906915879af3f62
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>