BUG-3200: remove UNMODIFIED nodes when preparing a candidate
A merge/write combination can result in an UNMODIFIED node being present
in the candidate. This patch adds a prepare-time pass to remove such
nodes, improving accuracy of the actual delta being applied.

This comes at the cost of changing the API contract of
DataTreeTip.prepare(): previously it was possible to pass the same
DataTreeModification multiple times, now each modification can be
prepared at most once. This should not pose a problem, since a
modification is required to be sealed, so we do not really have a use
case which would call for this.

Change-Id: I8396008ce33a95eead1142cae5e8de2c63124568
Signed-off-by: Jan Hajnar <jhajnar@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>