Update patch set 1
Patch Set 1:

I did some testing with putting it in <pluginManagement> section and it appears that the <reporting> section does indeed pull values from <pluginManagement> section. I'll submit a patch to odlparent to put it there when I get back to my office.

A question I have though is I noticed odlparent already provides a <reporting> section with the exact same values as yangtools is doing here. Can we remove the <reporting> section in yang-maven-plugin since it is inheriting the same values from odlparent anyway?

Patch-set: 1