BUG-4158: improve (Ordered)LeafSetNode memory footprint
LeafSetNodes have the interesting property that the child's identifier
already contains the child's value. We can take advantage of this fact
by not storing an entire map, but rather just the set of keys. We can
then instantiate LeafSetEntryNodes on-demand, lowering the persistent
memory footprint.

We create two classes for Mutable/Immutable view of this set, which can
be instantiated from each other. The immutable view uses an internal
ImmutableSet. The mutable view may inherit it and when it does it
attempts to retain it as long as no real modification is done.

Change-Id: I9462d757446bdf3c827b53a8e354b4436f794f42
Signed-off-by: Robert Varga <rovarga@cisco.com>