BUG-4158: introduce ImmutableOffsetMap
ImmutableOffsetMap stores its children in an array and shares the
key-to-offset map with other instances. In order to facilitate the
modification cycle, there is also a MutableOffsetMap, which uses a
LinkedHashMap to record any additions.

The conversion between the two reuses the backing array as much as
possible, to make transitions between modification cycles when no
structural modification occurs efficient.

Change-Id: I100469426c1cf741dd5545c5d11fd29aa3e4ab8b
Signed-off-by: Robert Varga <rovarga@cisco.com>