BUG-4275: Improve MapAdaptor and LazyContainerNode map overhead
Analysis of a heap dump has shown that for two-entry containers we end
up allocating HashMaps with default load factor. This boils down to
HashMap's copy constructor enforcing allocation to at least 16 entries
and also us not properly sizing for small maps.

MapAdaptor is taught to use HashMap.clone() and fall back to conservative
sizing on small inputs.

LazyContainerNode is taught to properly size the copy map when being
converted to a Mutable.

Change-Id: I6ad36aea31aab2e5e70176143d74e97c6d6a9ca5
Signed-off-by: Robert Varga <rovarga@cisco.com>