Bug 4145: Issue augmenting choice inside a list
The problem occurs when target node of uses-augment statement
is added by uses from another grouping and the uses statement
is not directly inside the grouping, but it is nested below
other elements (e.g. list, container etc.) in the grouping.
In this case the original of the target node is not found and
the IllegalStateException is thrown.

Change-Id: I51416b84e2bec81df814f214d2cebfaf93304d79
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>