BUG-4300: Make Lazy*ContainerNode really lazy
Our LazyContainerNode/MaterializedContainerNode combo meant that a
TreeNode was either non-materialized (Lazy) or fully-materialized (after
it was touched again). This meant we could end up stalling for
significant amount of time (when a large map was materialized), or ended
up materializing leaves which can never be modified -- like key leaves
in a list.

This patch fixes the problem by introducing a middle stage, where only
children which were modified are tracked and requests to non-modified
children are served from the data portion. This means materialization
occurs gradually as needed -- preventing stalls and improving memory
efficiency.

Change-Id: I480f675423f8e1f0e999ad609e40093e3f4a267e
Signed-off-by: Robert Varga <rovarga@cisco.com>