Update patch set 17
Patch Set 17:

> integration-patch-test-beryllium/36/ : FAILURE

Test results are almost the same as in current beryllium snapshot, but additionally there is Bug 4337 and Bug 4338.

There are of "Node was created by other transaction." type. I have expected this change to cause missing nodes, not additional nodes; so perhaps there is still something to fix in this change.

Patch-set: 17
Label: Verified=0