Update patch set 8
Patch Set 8:

Yes I agree, but JMXGenerator in controller expects QName as argument of unknownStatement in case it is possible to parse it. If argument of unknownStatement is string only, controller build will fail and rework of JMXGenerator will be required. Therefore, I propose this solution and the QName is created when it is possible to parse it (and the argument of unknownStatement is accessible as both QName and String) and in other cases the argument is accessible as string only.

Patch-set: 8
Reviewer: Gerrit User 2745 <2745@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0