Share instances of built-in types
There is simply no reason to instantiate the base type for whenever it
is referenced. Create a bunch of singletons which will be used when this
happens.

In order to achieve that, the type handling is reworked so that the
type/typedef statements reflect their relationship and also the
resulting TypeDefinition#getBaseType() reflects the typedef hierarchy
without intermittent 'type' statements interfering.

Change-Id: I2ce7d774e1bb08a4482b1f96d92c577a3668be78
Signed-off-by: Robert Varga <rovarga@cisco.com>