Update patch set 3
Patch Set 3:

So the reason for logging on debug comes from the controller, where everyone felt that these exceptions should not be logged by default -- which boils down to providing a modular Karaf logging configuration, which has not been designed/implemented.

The second problem is with Karaf's logger configuration/implementation, as you described -- that should really be investigated and fixed.

I really do not like changing the errors we emit based on the logger's configuration -- that leads to surprise when the user changes to debug and the error message they have been seeing no longer appears.

Patch-set: 3
Label: Verified=0