Update patch set 4
Patch Set 4:

(2 comments)

In my patch the error message still appears with debug enabled, just with more info. You've modified it to what I was trying to avoid - long, unfriendly, chained stack traces with a lot of noise that make it hard to find the relevant, user-facing error. I was trying to only print a  friendly/user-facing message by default, not only for karaf, but for the maven plugin where most errors will be emitted. Try running the maven plugin on a yang with an error - the errors traces look pretty ugly. The user really doesn't care about internals like the "SOURCE_LINKAGE phase failed" - they care that, e.g., "Imported module XXX was not found".  As an analogy, wouldn't it be annoying if the Java compiler printed long error traces with cryptic, internal info  emitted for a code compilation error?

Patch-set: 4
Label: Verified=0