Update patch set 22
Patch Set 22:

What happens if a non-presence container is explicitly created, and is therefore empty until a later write adds a child?  Is there a way that the child write could fail for missing parent?  With this behavior, it will be possible to write an empty container, then read to see that it's there ... and find it's gone ... not a pleasant experience.

Patch-set: 22
Label: Verified=0