Update patch set 22
Patch Set 22:

A write of non-presence container will be turned into a no-op, as it will appear and disappear in the same transaction.

Attempting to create a child will not fail, as the container will be created as needed.

I am not sure I follow what the use case would be for writing an empty non-presence container and expecting it to be there.

Patch-set: 22
Label: Verified=0