BUG-4355: mandatory node presence enforcement
Introduce mandatory leaf validation as specified in RFC6020. Mandatory
node presence is enforced whenever a presence container, a list entry,
and a case node.

The presence of mandatory leaves which do not have any ancestors other
than structural containers are not enforced as mandatory. The reason for
this is that the first transaction on such a DataTree would have to
introduce all mandatory leaves, as their non-presence would make other
transactions fail consistency checks.

Change-Id: I19df441026854955fd37f6e1ff7b4ea3c48c60b8
Signed-off-by: Robert Varga <rovarga@cisco.com>