Update patch set 7
Patch Set 7:

I compared a missing import error with the old parser (Li) and the new parser. The old parser resulted in the "Imported module ... not found" message to be printed in the maven "Failed to execute goal..." summary, which is ideally what you'd expect/want. So the new parser introduced a regression in that regard (which patch set #7 fixes). The old parser also printed the yang file line # in the message which is nice (another regression but beyond the scope of this patch).

Patch-set: 7