BUG-4629: cleanup YinSchemaSource
This patch introduces Yang/YinSchemaSourceRepresentation, which mark
YANG and YIN schema sources irrespective of the format in which they are
available.

It then adds YimXmlSchemaSource, which allows access to YIN
representation as a javax.xml.transform.Source. Finally it deprecates
YinSchemaSource and provides a utility YinDomSchemaSource, which can
translate a YinXmlSchemaSource in the DOM representation.

Change-Id: I832b2f14637eeb3183c6fe3254f8129bccc0a7ba
Signed-off-by: Robert Varga <robert.varga@pantheon.sk>