Bug 5019: Add QName param to NormalizedNodeWriter#leafSetEntryNode
NormalizedNodeWriter allows us to emit a single LeafSetEntryNode, which
breaks assumptions in NormalizedNodeStreamWriter -- specifically that
the receiver has seen its parent LeafsetNode.

In this case this is not true, so we need to supply the QName to pass
down the identity of the entry.

Change-Id: I45187800c62ec36b276724d230a922634c556a9b
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>
(cherry picked from commit d9193a3ec0a76392097f04d14d0547a7c00d5d27)