Update patch set 1
Patch Set 1:

Still in progress, some tests are failing due to changed lifecycle of list nodes in DataTree. Will be fixed in next patch set, but the question is: do we want to change the behavior for auto-deletes also in stable/beryllium ? or keep what it used to be for lists and choices.

Patch-set: 1
Reviewer: Gerrit User 84 <84@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0