Bug-5382 Structural lifecycle for lists and choices
In addition to structural lifecycle management for container nodes,
keyed lists(MapNode, MapEntryNode)and choices(ChoiceNode) are now
managed in the same way.

This changes the behavior of (Ordered)MapNodes and ChoiceCaseNodes
inside DataTree. They are no longer kept empty and are created on
the fly when child node is submitted.

Change-Id: I68220043c0f335324e5a8788b2872f57b750aae9
Signed-off-by: Maros Marsalek <mmarsale@cisco.com>