Update patch set 3
Patch Set 3:

(3 comments)

overall, I think explicit tests of equals() should be invoking the method rather than rely on assertEquals(), since we want to be in control on which object that method is invoked.

Also when converting to assertEquals/assertNotEquals in other places, the order of arguments is important to convey what the expected value is -- there is confusion in quite a few places about that.

Patch-set: 3
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0