FilesystemSchemaSourceCache reduce sourceIdToFile(...) visibility
Change sourceIdToFile(...) scope to default.  public is not needed, as nothing
calls this function from outside the package.  I pulled autorelease, and
confirmed that no other projects call this function.  Some tests within the
package test this method explicitly, and there is no easy way to decrase
visibility to private.  The @Beta annotation was removed as this API is no
longer advertised publically.

Change-Id: Iacb41bea5eef7477b279f21b0fe2a103dcce3e4b
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>