Address FIXME for QueuedNotificationManager
The QueuedNotificationManager was originally implemented to loop forever in
attempt to offer the target notification queue a notification.  However, with a
reasonably large queue capacity and long timeout, if queueing is still
unsuccessful, then the listener is likely in an unrecoverable state (deadlock
or endless loop).  To avoid one rogue listener bringing everyone to a halt, the
number of tries is limited to 10 (10 minutes), and an appropriate warning
message is printed if the offer attempt fails.

Change-Id: Ie01ee475cfdda2c93c346d3b7ea823ccfc609b83
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>