Replace tabs in java files with spaces
A tab may be interpreted as a different number of columns depending on
environment.  On the hand, a space is always interpreted as just one column.
This change replaces all instances of tab characters with 4 spaces within
java files.

The following UNIX command was run in order to automate replacement:

> find . -type f -name "*.java" -exec sed -i 's/\t/    /g' {} \;

An appropriate comment was added to CheckCodingStyleTestClass.java to indicate
that the class contains intentional checkstyle errors (i.e., tabs) to test
the efficacy of checkstyle rules.

Change-Id: Ie51eadda2169d1bb15c8aef1ce01e625b9b9972d
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>