BUG-6652: generate features.xml with karaf-maven-plugin [WIP]
With Karaf 3.0.8, this produces a near-equivalent features.xml to the
previous, statically-constructed version. (The exception for the time
being is the inclusion of wrapped POM-type dependencies corresponding
to the features.)

This patch explicitly scopes test dependencies and removes those which
are provided in the compile scope (by features the module depends on).

Change-Id: I758da5c7f2b9a1b54b49d122a408a53eea22927c
Signed-off-by: Stephen Kitt <skitt@redhat.com>