Remove @Deprecated CompositeObjectRegistration
Not sure what ODL policy is re. clean up of deprecated classes, so I'm
proposing to killing one, and see what happens... ;-)
https://twitter.com/drdeprecator

If this must be kept, then I'll fix the catch (Exception e)
e.printStackTrace() in it, with LOGGER, because this is the only
critical issue yangtools has on Sonar.

Change-Id: Ice796b5cae5ce54df88e3227f9ec045e115c921b
Signed-off-by: Michael Vorburger <vorburger@redhat.com>