Introduce WritableObject/WritableIdentifier
When serializing nested Identifiers, it is sometimes useful
to be able to skip the object header in the stream. This patch
introduces WritableObject interface and WritableIdentifier
as the unification of WritableObject and WritableIdentifier.

It furthermore adds utilities to write long values if a format
which uses minimum number of bytes.

Change-Id: I8459e2883b4ab3dff113000cea43a452eacd3324
Signed-off-by: Robert Varga <rovarga@cisco.com>