Update patch set 1
Patch Set 1: Code-Review-1

I think this should be done via profiles. There is no point in activating the plugin at all. That is to say that both model-parent and config-parent (in controller) should be modified so that yang-maven-plugin is activated in a profile (say, processYang ,which is on by default.

The same effect can be then achieve by 'mvn -P!processYang'.

Patch-set: 1
Label: Code-Review=-1