[DO NOT MERGE] Drop yangtools-parent
This emerged from trying to use odlparent's bundle-parent for bundles,
and thus be able to remove the yangtools overrides for
maven-jar-plugin and maven-javadoc-plugin in particular.

yang-maven-plugin-it has a couple of hard-code version numbers which
need to be fixed. There are probably features this removes from
yangtools-parent which should be replicated elsewhere; in particular,
the Sonar/JaCoCo integration should be available for everyone.

Change-Id: Id41e61eb3da2f30be29981a11387bde467acae3d
Signed-off-by: Stephen Kitt <skitt@redhat.com>