Add QName.hashCode() cache
QName is usually involved in hash-based lookups, hence it is
useful to reuse a computed hash code so that constants do not
spend cycles on computing it more than once.

Add a field which will be lazily initiated to contain the result
of hashCode(). This has no impact on memory footprint.

Change-Id: I8a0b89599ed1029e2c572535826c52b4b837e265
Signed-off-by: Robert Varga <rovarga@cisco.com>