Update patch set 4
Patch Set 4:

I'm sorry, I should have told about this. We drove more tests, and the result of skipping yang generation is not always working correctly.

Particularly, files in META_INF are not copied to the final jar, and the yang service will fail. We are checking if there are more problems like this, and submit an appropriate patch when corrected.

I'm so sorry, until then, please, don't use this feature. Anyway, thank you very much for your attention to this problem. We keep on it.

Patch-set: 4
Reviewer: Gerrit User 5440 <5440@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0