Bug 6158: Optimization of if-feaures resolution during SchemaContext assembly
It is unnecessary to check both declared and effective substatements everytime, because
effective substatements are empty at the beginning and they are filled subsequently by
copying of declared. Therefore this patch check whether a feature is supported only for
declarted statements during their build and also creation of their copy.
In addition this patch introduces ALL_FEATURES predicate constant. The constant is passed
as argument into default SchemaContext build, and so we can simply return true and skip rest
of areFeaturesSupported method.
Moreover the result of areFeaturesSupported method is cached and so it is not computed
more times for any statement.

Change-Id: If8c87975712b65f9f28b640ac0b634576e08c191
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>