Minor refactoring in YangToSourcesPluginTestIT
Just to reduce a few lines of copy/paste.  I actually did this because I
wanted to call the factored out method for a real fix, but then ended up
not using it; so I'm putting in this proposed minor refactoring as a
first small low prio change, and the more important real fix as a
separate change coming up next.

Change-Id: Ic888310f1ae74e58eaa96e41c0cb53ab2215023e
Signed-off-by: Michael Vorburger <vorburger@redhat.com>