Update patch set 1
Patch Set 1:

Not sure what to do with the unit test. Also, SpecialCase.NOTNULL seem to be not used anymore, perhaps should be removed, with code that expects that? Finally, I am not sure what is MissingSubstatementException for, when InvalidSubstatementException is thrown when there is not enough substatements.

Patch-set: 1
Label: Verified=0