Structural lifecycle for lists and choices
In addition to structural lifecycle management for non-presence
container nodes, keyed lists(MapNode) and choices(ChoiceNode) are now
managed in the same way.

The behavior of these nodes is now different inside DataTree.
They are no longer kept empty and are created on the fly when child
node is submitted.

JIRA: YANGTOOLS-575
Change-Id: I68220043c0f335324e5a8788b2872f57b750aae9
Signed-off-by: Maros Marsalek <mmarsale@cisco.com>
Signed-off-by: Igor Foltin <ifoltin@cisco.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>