Update patch set 7
Patch Set 7:

(2 comments)

> The last finished run waited 125 seconds for Restconf to start working

That was also the time for builds around this one, but this patch achieved 95 seconds [1]. No idea why, but I would say performance impact is not visibly negative.

The test however does not conform to full RFC6020 section 6.1.3
so I would recommend to test in a string without indentation whitespace.

[1] https://jenkins.opendaylight.org/releng/job/yangtools-distribution-check-boron/232/

Patch-set: 7
Label: Verified=0