Update patch set 1
Patch Set 1: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/yangtools-verify-beryllium-mvn33-openjdk8/5/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-verify-beryllium-mvn33-openjdk8/5/ )

https://jenkins.opendaylight.org/releng/job/yangtools-verify-beryllium-mvn32-openjdk7/5/ : FAILURE

https://jenkins.opendaylight.org/releng/job/yangtools-validate-autorelease-beryllium/331/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-validate-autorelease-beryllium/331/ )

https://jenkins.opendaylight.org/releng/job/yangtools-distribution-check-beryllium/9/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-distribution-check-beryllium/9/ )

https://jenkins.opendaylight.org/releng/job/yangtools-verify-beryllium-mvn33-openjdk7/6/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-verify-beryllium-mvn33-openjdk7/6/ )

https://jenkins.opendaylight.org/releng/job/yangtools-verify-beryllium-mvn32-openjdk8/5/ : FAILURE

Patch-set: 1