Update patch set 8
Patch Set 8:

> Just FYI, you can reduce the load on the infrastructure and save
 > time by using specific code-words to restart only failed tests (in
 > this case, test-distro with distro spelt in full). See
 > https://nexus.opendaylight.org/content/sites/site/org.opendaylight.releng.builder/master/jenkins.html#jjb_templates
 > for the complete list...

Noted.  Question: will this add the +1 from jenkins-releng, like will it remember that 2 passed and 1 failed previously, and now the failed third passed, so that it can +1 it?  It's because I've noticed that reviewers typically just ignore (rightfully IMHO) changes that don't have the Verified from jenkins-releng.

Patch-set: 8