Update patch set 6
Patch Set 6: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/yangtools-verify-carbon-mvn32-openjdk8/58/ : FAILURE

https://jenkins.opendaylight.org/releng/job/yangtools-verify-carbon-mvn33-openjdk8/58/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-verify-carbon-mvn33-openjdk8/58/ )

https://jenkins.opendaylight.org/releng/job/yangtools-validate-autorelease-carbon/8/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-validate-autorelease-carbon/8/ )

https://jenkins.opendaylight.org/releng/job/yangtools-distribution-check-carbon/26/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-distribution-check-carbon/26/ )

Patch-set: 6
Label: Verified=-1