Update patch set 1
Patch Set 1:

When we provide input like "1.0", the method valueOf calls the method create which adds the patch number if it is not defined.
So eventually, the SemVer object will have valid form 1.0.0.
The problem here is that in this case, valueOf throws a NumberFormatException. This patch fixes that.

Patch-set: 1
Reviewer: Gerrit User 4561 <4561@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0