Update patch set 9
Patch Set 9: Verified-1

Build Failed 

https://jenkins.opendaylight.org/releng/job/yangtools-verify-carbon-mvn32-openjdk8/75/ : FAILURE

https://jenkins.opendaylight.org/releng/job/yangtools-distribution-check-carbon/42/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-distribution-check-carbon/42/ )

https://jenkins.opendaylight.org/releng/job/yangtools-verify-carbon-mvn33-openjdk8/75/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-verify-carbon-mvn33-openjdk8/75/ )

https://jenkins.opendaylight.org/releng/job/yangtools-validate-autorelease-carbon/13/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch (eg. no Beryllium version inside of Boron branch) ( https://jenkins.opendaylight.org/releng//job/yangtools-validate-autorelease-carbon/13/ )

Patch-set: 9
Label: Verified=-1